Moved test decorator to separate file
